+++ src/main/java/edu/hm/hafner/analysis/parser/IdeaInspectionParser.java
@@ -3,6 +3,7 @@ package edu.hm.hafner.analysis.parser;
 import java.util.List;
 import java.util.Optional;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.StringEscapeUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -23,6 +24,7 @@ import edu.hm.hafner.util.XmlElementUtil;
  */
 public class IdeaInspectionParser extends IssueParser {
     private static final long serialVersionUID = 3307389086106375473L;
+    private static final String PATH_PREFIX = "file://";
 
     @Override
     public Report parse(final ReaderFactory readerFactory) throws ParsingException {
@@ -39,7 +41,7 @@ public class IdeaInspectionParser extends IssueParser {
             Optional<Element> problemClass = XmlElementUtil.getFirstChildElementByName(element, "problem_class");
             if (problemClass.isPresent()) {
                 Element problem = problemClass.get();
-                IssueBuilder builder = new IssueBuilder().setFileName(file)
+                IssueBuilder builder = new IssueBuilder().setFileName(stripPathPrefix(file))
                         .setLineStart(Integer.parseInt(getChildValue(element, "line")))
                         .setCategory(StringEscapeUtils.unescapeXml(getValue(problem)))
                         .setMessage(StringEscapeUtils.unescapeXml(getChildValue(element, "description")))
@@ -59,6 +61,10 @@ public class IdeaInspectionParser extends IssueParser {
             priority = Severity.WARNING_HIGH;
         }
         return priority;
+    }
+
+    private String stripPathPrefix(final String file) {
+        return StringUtils.removeStart(file, PATH_PREFIX);
     }
 
     private String getValue(final Element element) {